pyproject.toml
[project]
name = "mlops-coffee-match"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.68.0",
    "openai>=1.108.0",
    "pandas>=2.3.2",
    "pandas-stubs>=2.3.2.250827",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "shortuuid>=1.0.13",
]
